Contextos antigos removidos. Uso do migrations correto após 3ª tentativa.

+++ TempusDigitalMVC/Views/Home/RelatorioDados.cs
@@ -9,13 +9,13 @@ namespace TempusDigitalMVC.Views.Home
     {
         //https://stackoverflow.com/questions/11296634/pass-a-simple-string-from-controller-to-a-view-mvc3
 
-        public double MediaRendaFamiliar { get; set; }
+        public decimal MediaRendaFamiliar { get; set; }
         public int QtdePessoaRendaAcimaMedia { get; set; }
         public int QtdeClasseA { get; set; }
         public int QtdeClasseB { get; set; }
         public int QtdeClasseC { get; set; }
-        public double RendaMes { get; set; }
-        public double RendaSemana { get; set; }
-        public double RendaHoje { get; set; }
+        public decimal RendaMes { get; set; }
+        public decimal RendaSemana { get; set; }
+        public decimal RendaHoje { get; set; }
     }
 }
